move MineMenu Integration to it's own class
