+++ src/main/java/de/mari_023/fabric/ae2wtlib/wut/ItemWUT.java
@@ -3,7 +3,7 @@ package de.mari_023.fabric.ae2wtlib.wut;
 import appeng.container.ContainerLocator;
 import appeng.core.AEConfig;
 import de.mari_023.fabric.ae2wtlib.ae2wtlib;
-import de.mari_023.fabric.ae2wtlib.client.ae2wtlibclient;
+import de.mari_023.fabric.ae2wtlib.client.MineMenuIntegration;
 import de.mari_023.fabric.ae2wtlib.terminal.IInfinityBoosterCardHolder;
 import de.mari_023.fabric.ae2wtlib.terminal.ItemWT;
 import net.fabricmc.api.EnvType;
@@ -32,7 +32,7 @@ public class ItemWUT extends ItemWT implements IInfinityBoosterCardHolder {
     @Override
     public TypedActionResult<ItemStack> use(final World w, final PlayerEntity player, final Hand hand) {
         if(player.isSneaking()) {
-            if(w.isClient()) ae2wtlibclient.openMineMenu();
+            if(w.isClient()) MineMenuIntegration.openMineMenu();
         } else openWirelessTerminalGui(player.getStackInHand(hand), player, hand);
         return new TypedActionResult<>(ActionResult.SUCCESS, player.getStackInHand(hand));
     }
